change all double to float

+++ include/lshape_estimator.hpp
@@ -14,12 +14,12 @@ private:
 public:
   orientation_calc(std::string value) : criterion_(value){};
   ~orientation_calc(){};
-  bool LshapeFitting(const pcl::PointCloud<pcl::PointXYZI> &cluster, double &theta_optim);
-  double calcClosenessCriterion(const std::vector<double> &C_1, const std::vector<double> &C_2);
-  double calc_var(const std::vector<double> &v);
-  double calc_variances_criterion(const std::vector<double> &C_1, const std::vector<double> &C_2);
-  double calc_closeness_criterion(const std::vector<double> &C_1, const std::vector<double> &C_2);
-  double calc_area_criterion(const std::vector<double> &C_1, const std::vector<double> &C_2);
+  bool LshapeFitting(const pcl::PointCloud<pcl::PointXYZI> &cluster, float &theta_optim);
+  float calcClosenessCriterion(const std::vector<float> &C_1, const std::vector<float> &C_2);
+  float calc_var(const std::vector<float> &v);
+  float calc_variances_criterion(const std::vector<float> &C_1, const std::vector<float> &C_2);
+  float calc_closeness_criterion(const std::vector<float> &C_1, const std::vector<float> &C_2);
+  float calc_area_criterion(const std::vector<float> &C_1, const std::vector<float> &C_2);
 };
